Just one minor quality of life changes
I just thinking about minding my own business in the backend while leaving the UI UX decisions to the frontend because it wasn't my thing. So I'll stop stressing about my choice in UI UX and just go along with it. Maybe I'll get myself a partner to build websites together with me in the Backend.